+++ crates/sqlite-store/src/account.rs
@@ -14,12 +14,13 @@ use miden_client::account::{
     AccountId,
     AccountIdPrefix,
     AccountStorage,
+    Address,
     StorageMap,
     StorageSlot,
     StorageSlotType,
 };
-use miden_client::asset::{Asset, AssetVault, FungibleAsset, NonFungibleDeltaAction};
-use miden_client::crypto::{MerklePath, MerkleStore, SmtLeaf, SmtProof};
+use miden_client::asset::{Asset, AssetVault, AssetWitness, FungibleAsset, NonFungibleDeltaAction};
+use miden_client::crypto::{MerkleStore, SmtLeaf, SmtProof};
 use miden_client::store::{AccountRecord, AccountStatus, StoreError};
 use miden_client::utils::{Deserializable, Serializable};
 use miden_client::{AccountError, Felt, Word};
@@ -128,17 +129,24 @@ impl SqliteStore {
             return Ok(None);
         };
 
-        Ok(Some(AccountRecord::new(
-            Account::from_parts(header.id(), vault, storage, account_code, header.nonce()),
-            status,
-        )))
+        let account = Account::new_unchecked(
+            header.id(),
+            vault,
+            storage,
+            account_code,
+            header.nonce(),
+            status.seed().copied(),
+        );
+
+        let addresses = query_account_addresses(conn, header.id())?;
+        Ok(Some(AccountRecord::new(account, status, addresses)))
     }
 
     pub(crate) fn insert_account(
         conn: &mut Connection,
         merkle_store: &Arc<RwLock<MerkleStore>>,
         account: &Account,
-        account_seed: Option<Word>,
+        initial_address: &Address,
     ) -> Result<(), StoreError> {
         let tx = conn.transaction().into_store_error()?;
 
@@ -151,7 +159,9 @@ impl SqliteStore {
         )?;
 
         Self::insert_assets(&tx, account.vault().root(), account.vault().assets())?;
-        Self::insert_account_header(&tx, &account.into(), account_seed)?;
+        Self::insert_account_header(&tx, &account.into(), account.seed())?;
+
+        Self::insert_address(&tx, initial_address, account.id())?;
 
         tx.commit().into_store_error()?;
 
@@ -284,7 +294,7 @@ impl SqliteStore {
         merkle_store: &Arc<RwLock<MerkleStore>>,
         account_id: AccountId,
         faucet_id_prefix: AccountIdPrefix,
-    ) -> Result<Option<(Asset, MerklePath)>, StoreError> {
+    ) -> Result<Option<(Asset, AssetWitness)>, StoreError> {
         let header = Self::get_account_header(conn, account_id)?
             .ok_or(StoreError::AccountDataNotFound(account_id))?
             .0;
@@ -301,9 +311,10 @@ impl SqliteStore {
 
         let merkle_store = merkle_store.read().expect("merkle_store read lock not poisoned");
 
-        let merkle_path = get_asset_proof(&merkle_store, header.vault_root(), &asset)?;
+        let proof = get_asset_proof(&merkle_store, header.vault_root(), &asset)?;
+        let witness = AssetWitness::new(proof)?;
 
-        Ok(Some((asset, merkle_path)))
+        Ok(Some((asset, witness)))
     }
 
     /// Retrieves a specific item from the account's storage map without loading the entire storage.
@@ -331,14 +342,23 @@ impl SqliteStore {
         };
 
         let item = map.get(&key);
-
         let merkle_store = merkle_store.read().expect("merkle_store read lock not poisoned");
 
-        let (value, path) = get_storage_map_item_proof(&merkle_store, map.root(), key)?;
-        let leaf = SmtLeaf::new_single(key, value);
+        // TODO: change the api of get_storage_map_item_proof
+        let path = get_storage_map_item_proof(&merkle_store, map.root(), key)?.1;
+        let leaf = SmtLeaf::new_single(StorageMap::hash_key(key), item);
         let proof = SmtProof::new(path, leaf)?;
 
-        Ok((item, StorageMapWitness::new(proof)))
+        let witness = StorageMapWitness::new(proof, [key])?;
+
+        Ok((item, witness))
+    }
+
+    pub(crate) fn get_account_addresses(
+        conn: &mut Connection,
+        account_id: AccountId,
+    ) -> Result<Vec<Address>, StoreError> {
+        query_account_addresses(conn, account_id)
     }
 
     // ACCOUNT DELTA HELPERS
@@ -786,6 +806,19 @@ impl SqliteStore {
 
         Ok(())
     }
+
+    fn insert_address(
+        tx: &Transaction<'_>,
+        address: &Address,
+        account_id: AccountId,
+    ) -> Result<(), StoreError> {
+        const QUERY: &str = insert_sql!(addresses { address, account_id } | REPLACE);
+        let serialized_address = address.to_bytes();
+        tx.execute(QUERY, params![serialized_address, account_id.to_hex(),])
+            .into_store_error()?;
+
+        Ok(())
+    }
 }
 
 // HELPERS
@@ -982,6 +1015,28 @@ fn query_account_headers(
         .collect::<Result<Vec<(AccountHeader, AccountStatus)>, StoreError>>()
 }
 
+fn query_account_addresses(
+    conn: &Connection,
+    account_id: AccountId,
+) -> Result<Vec<Address>, StoreError> {
+    const ADDRESS_QUERY: &str = "SELECT address FROM addresses";
+
+    let query = format!("{ADDRESS_QUERY} WHERE ACCOUNT_ID = '{}'", account_id.to_hex());
+    conn.prepare(&query)
+        .into_store_error()?
+        .query_map([], |row| {
+            let address: Vec<u8> = row.get(0)?;
+            Ok(address)
+        })
+        .into_store_error()?
+        .map(|result| {
+            let serialized_address = result.into_store_error()?;
+            let address = Address::read_from_bytes(&serialized_address)?;
+            Ok(address)
+        })
+        .collect::<Result<Vec<Address>, StoreError>>()
+}
+
 #[cfg(test)]
 mod tests {
     use std::collections::BTreeMap;
@@ -996,7 +1051,10 @@ mod tests {
         AccountDelta,
         AccountHeader,
         AccountId,
+        AccountIdAddress,
         AccountType,
+        Address,
+        AddressInterface,
         StorageMap,
         StorageSlot,
     };
@@ -1084,12 +1142,15 @@ mod tests {
         .with_supports_all_types();
 
         // Create and insert an account
-        let (account, seed) = AccountBuilder::new([0; 32])
+        let account = AccountBuilder::new([0; 32])
             .account_type(AccountType::RegularAccountImmutableCode)
             .with_auth_component(AuthRpoFalcon512::new(PublicKey::new(EMPTY_WORD)))
             .with_component(dummy_component)
             .build()?;
-        store.insert_account(&account, Some(seed)).await?;
+
+        let default_address =
+            Address::AccountId(AccountIdAddress::new(account.id(), AddressInterface::Unspecified));
+        store.insert_account(&account, default_address).await?;
 
         let mut storage_delta = AccountStorageDelta::new();
         storage_delta.set_item(1, [ZERO, ZERO, ZERO, ONE].into());
@@ -1177,7 +1238,9 @@ mod tests {
             .with_component(dummy_component)
             .with_assets(assets.clone())
             .build_existing()?;
-        store.insert_account(&account, None).await?;
+        let default_address =
+            Address::AccountId(AccountIdAddress::new(account.id(), AddressInterface::Unspecified));
+        store.insert_account(&account, default_address).await?;
 
         let mut storage_delta = AccountStorageDelta::new();
         storage_delta.set_item(1, EMPTY_WORD);