+++ crates/web-client/src/export.rs
@@ -1,9 +1,10 @@
 use miden_client::Word;
-use miden_client::account::AccountFile;
+use miden_client::account::AccountFile as NativeAccountFile;
 use miden_client::store::NoteExportType;
 use miden_client::utils::{Serializable, get_public_keys_from_account};
 use wasm_bindgen::prelude::*;
 
+use crate::models::account_file::AccountFile;
 use crate::models::account_id::AccountId;
 use crate::{WebClient, js_error_with_context};
 
@@ -69,7 +70,10 @@ impl WebClient {
     }
 
     #[wasm_bindgen(js_name = "exportAccountFile")]
-    pub async fn export_account_file(&mut self, account_id: AccountId) -> Result<JsValue, JsValue> {
+    pub async fn export_account_file(
+        &mut self,
+        account_id: AccountId,
+    ) -> Result<AccountFile, JsValue> {
         if let Some(client) = self.get_mut_inner() {
             let account = client
                 .get_account(account_id.into())
@@ -86,7 +90,6 @@ impl WebClient {
                 .ok_or(JsValue::from_str("No account found"))?;
 
             let keystore = self.keystore.clone().expect("Keystore not initialized");
-            let account_seed = account.seed().copied();
             let account = account.into();
 
             let mut key_pairs = vec![];
@@ -103,13 +106,9 @@ impl WebClient {
                 );
             }
 
-            let account_data = AccountFile::new(account, account_seed, key_pairs);
+            let account_data = NativeAccountFile::new(account, key_pairs);
 
-            let serialized_input_note_bytes =
-                serde_wasm_bindgen::to_value(&account_data.to_bytes())
-                    .map_err(|_| JsValue::from_str("Serialization error"))?;
-
-            Ok(serialized_input_note_bytes)
+            Ok(AccountFile::from(account_data))
         } else {
             Err(JsValue::from_str("Client not initialized"))
         }